# Speedup ActiveModel::Type::Integer#serialize by up to 10x